[mod/#241] CheckMaxLength -> data class로 변경 및 파일 위치 이동
